refactor(pr): Extract labels-related code to separate file
