+++ lib/workers/repository/update/pr/labels.ts
@@ -1,0 +1,12 @@
+import is from '@sindresorhus/is';
+import type { RenovateConfig } from '../../../../config/types';
+import * as template from '../../../../util/template';
+
+export function prepareLabels(config: RenovateConfig): string[] {
+  const labels = config.labels ?? [];
+  const addLabels = config.addLabels ?? [];
+  return [...new Set([...labels, ...addLabels])]
+    .filter(is.nonEmptyStringAndNotWhitespace)
+    .map((label) => template.compile(label, config))
+    .filter(is.nonEmptyStringAndNotWhitespace);
+}
